pyproject.toml
[tool.poetry]
name = "sevro"
version = "0.1.0"
description = ""
authors = ["Matt Lewis <matt@mplewis.com>"]

[tool.poetry.dependencies]
python = "^3.8"
pyyaml = "^5.3.1"
schematics = "^2.1.0"
colored = "^1.4.2"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
ipython = "^7.16.1"
black = "^19.10b0"
flake8 = "^3.8.3"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
